refactor: admin 도메인 패키지 변경 (#133)
