+++ src/main/java/site/gachontable/domain/admin/exception/AdminNotFoundException.java
@@ -1,0 +1,10 @@
+package site.gachontable.domain.admin.exception;
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class AdminNotFoundException extends ServiceException {
+    public AdminNotFoundException() {
+        super(ErrorCode.ADMIN_NOT_FOUND);
+    }
+}
