+++ src/main/java/site/gachontable/domain/admin/port/in/AdminLogin.java
@@ -1,0 +1,8 @@
+package site.gachontable.domain.admin.port.in;
+
+import site.gachontable.presentation.admin.dto.response.AdminLoginResponse;
+
+public interface AdminLogin {
+
+    AdminLoginResponse execute(String id, String password);
+}
