+++ src/main/java/site/gachontable/domain/admin/port/in/GetSeatings.java
@@ -1,0 +1,9 @@
+package site.gachontable.domain.admin.port.in;
+
+import site.gachontable.presentation.admin.dto.response.SeatingsResponse;
+import site.gachontable.infra.security.principal.AuthDetails;
+
+public interface GetSeatings {
+
+    SeatingsResponse execute(AuthDetails authDetails);
+}
