+++ src/main/java/site/gachontable/domain/admin/port/in/ManagePub.java
@@ -1,0 +1,9 @@
+package site.gachontable.domain.admin.port.in;
+
+import site.gachontable.infra.security.principal.AuthDetails;
+import site.gachontable.presentation.admin.dto.request.PubManageRequest;
+
+public interface ManagePub {
+
+    String execute(AuthDetails authDetails, PubManageRequest request);
+}
