+++ src/main/java/site/gachontable/domain/admin/port/in/UpdateStatus.java
@@ -1,0 +1,12 @@
+package site.gachontable.domain.admin.port.in;
+
+import site.gachontable.presentation.admin.dto.request.UpdateStatusRequest;
+import site.gachontable.infra.security.principal.AuthDetails;
+import site.gachontable.domain.shared.dto.response.RegisterResponse;
+
+public interface UpdateStatus {
+
+    RegisterResponse executeForOpenStatus(AuthDetails authDetails, UpdateStatusRequest request);
+
+    RegisterResponse executeForWaitingStatus(AuthDetails authDetails, UpdateStatusRequest request);
+}
