+++ src/main/java/site/gachontable/domain/admin/service/GetSeatingsImpl.java
@@ -1,15 +1,16 @@
-package site.gachontable.gachontablebe.domain.admin.usecase;
+package site.gachontable.domain.admin.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import site.gachontable.gachontablebe.domain.admin.domain.repository.AdminRepository;
-import site.gachontable.gachontablebe.domain.admin.exception.AdminNotFoundException;
-import site.gachontable.gachontablebe.domain.admin.presentation.dto.response.SeatingsResponse;
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-import site.gachontable.gachontablebe.domain.pub.domain.Pub;
-import site.gachontable.gachontablebe.domain.seating.domain.Seating;
-import site.gachontable.gachontablebe.domain.seating.domain.respository.SeatingRepository;
+import site.gachontable.domain.admin.port.out.AdminRepository;
+import site.gachontable.domain.admin.exception.AdminNotFoundException;
+import site.gachontable.domain.admin.port.in.GetSeatings;
+import site.gachontable.presentation.admin.dto.response.SeatingsResponse;
+import site.gachontable.infra.security.principal.AuthDetails;
+import site.gachontable.domain.pub.domain.Pub;
+import site.gachontable.domain.seating.domain.Seating;
+import site.gachontable.domain.seating.port.out.SeatingRepository;
 
 import java.time.LocalDateTime;
