--- src/main/java/site/gachontable/gachontablebe/domain/admin/exception/AdminNotFoundException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class AdminNotFoundException extends ServiceException {
-    public AdminNotFoundException() {
-        super(ErrorCode.ADMIN_NOT_FOUND);
-    }
-}
