--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/CallUserRequest.java
@@ -1,6 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-import java.util.UUID;
-
-public record CallUserRequest(UUID waitingId) {
-}
