--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/ExitUserRequest.java
@@ -1,4 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-public record ExitUserRequest(Long seatingId) {
-}
