--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/UpdateStatusRequest.java
@@ -1,4 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-public record UpdateStatusRequest(Boolean status) {
-}
