--- src/main/java/site/gachontable/gachontablebe/domain/admin/usecase/AdminLogin.java
@@ -1,8 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.usecase;
-
-import site.gachontable.gachontablebe.domain.admin.presentation.dto.response.AdminLoginResponse;
-
-public interface AdminLogin {
-
-    AdminLoginResponse execute(String id, String password);
-}
