--- src/main/java/site/gachontable/gachontablebe/domain/admin/usecase/GetSeatings.java
@@ -1,9 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.usecase;
-
-import site.gachontable.gachontablebe.domain.admin.presentation.dto.response.SeatingsResponse;
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-
-public interface GetSeatings {
-
-    SeatingsResponse execute(AuthDetails authDetails);
-}
